requirements.txt
flask==0.11
flask-script==2.0.5
flask-mongoengine==0.8.2
py-bcrypt==0.4
jsonschema==2.5.1
pika==1.1.0
requests==2.22.0
attrdict==2.0.1
pytest==5.0.1
pyyaml==5.1.1
hiyapyco==0.4.14
pykwalify==1.7.0
falcon==2.0.0
falcon-jsonify==1.2
flask-restful==0.3.7
flask-restplus==0.12.1
flask-wtf==0.14.2
werkzeug==0.15.5
gunicorn==19.9.0
ecmwf-api-client==1.5.4
pymongo==3.10.1
pymemcache==3.0.0
redis==3.4.1
markdown==3.2.1
minio==5.0.7
tiny-kubernetes==1.1.1
deepmerge==0.1.0
flask-swagger-ui==3.25.0
ldap3==2.7
docker==4.2.0
python-keycloak==0.24.0